+++ nerf/model/__init__.py
@@ -1,0 +1,2 @@
+from nerf.model.nerf import NeRF
+from nerf.model.model_helper import *
