+++ nerf/model/model_helper.py
@@ -1,0 +1,192 @@
+import os
+import torch
+import torch.nn as nn
+import numpy as np
+
+from nerf.model.nerf import NeRF
+
+
+# Positional encoding (section 5.1)
+class Embedder:
+    def __init__(self, **kwargs):
+        self.kwargs = kwargs
+        self.create_embedding_fn()
+
+    def create_embedding_fn(self):
+        embed_fns = []
+        d = self.kwargs["input_dims"]
+        out_dim = 0
+        if self.kwargs["include_input"]:
+            embed_fns.append(lambda x: x)
+            out_dim += d
+
+        max_freq = self.kwargs["max_freq_log2"]
+        N_freqs = self.kwargs["num_freqs"]
+
+        if self.kwargs["log_sampling"]:
+            freq_bands = 2.0 ** torch.linspace(0.0, max_freq, steps=N_freqs)
+        else:
+            freq_bands = torch.linspace(2.0**0.0, 2.0**max_freq, steps=N_freqs)
+
+        for freq in freq_bands:
+            for p_fn in self.kwargs["periodic_fns"]:
+                embed_fns.append(lambda x, p_fn=p_fn, freq=freq: p_fn(x * freq))
+                out_dim += d
+
+        self.embed_fns = embed_fns
+        self.out_dim = out_dim
+
+    def embed(self, inputs):
+        return torch.cat([fn(inputs) for fn in self.embed_fns], -1)
+
+
+def get_embedder(multires, i=0):
+    if i == -1:
+        return nn.Identity(), 3
+
+    embed_kwargs = {
+        "include_input": True,
+        "input_dims": 3,
+        "max_freq_log2": multires - 1,
+        "num_freqs": multires,
+        "log_sampling": True,
+        "periodic_fns": [torch.sin, torch.cos],
+    }
+
+    embedder_obj = Embedder(**embed_kwargs)
+    embed = lambda x, eo=embedder_obj: eo.embed(x)
+    return embed, embedder_obj.out_dim
+
+
+def batchify(fn, chunk):
+    """Constructs a version of 'fn' that applies to smaller batches."""
+    if chunk is None:
+        return fn
+
+    def ret(inputs):
+        return torch.cat(
+            [fn(inputs[i : i + chunk]) for i in range(0, inputs.shape[0], chunk)], 0
+        )
+
+    return ret
+
+
+def run_network(inputs, viewdirs, fn, embed_fn, embeddirs_fn, netchunk=1024 * 64):
+    """Prepares inputs and applies network 'fn'."""
+    inputs_flat = torch.reshape(inputs, [-1, inputs.shape[-1]])
+    embedded = embed_fn(inputs_flat)
+
+    if viewdirs is not None:
+        input_dirs = viewdirs[:, None].expand(inputs.shape)
+        input_dirs_flat = torch.reshape(input_dirs, [-1, input_dirs.shape[-1]])
+        embedded_dirs = embeddirs_fn(input_dirs_flat)
+        embedded = torch.cat([embedded, embedded_dirs], -1)
+
+    outputs_flat = batchify(fn, netchunk)(embedded)
+    outputs = torch.reshape(
+        outputs_flat, list(inputs.shape[:-1]) + [outputs_flat.shape[-1]]
+    )
+    return outputs
+
+
+def create_nerf(args, device):
+    """Instantiate NeRF's MLP model."""
+    embed_fn, input_ch = get_embedder(args.multires, args.i_embed)
+
+    input_ch_views = 0
+    embeddirs_fn = None
+    if args.use_viewdirs:
+        embeddirs_fn, input_ch_views = get_embedder(args.multires_views, args.i_embed)
+    output_ch = 5 if args.N_importance > 0 else 4
+    skips = [4]
+    model = NeRF(
+        D=args.netdepth,
+        W=args.netwidth,
+        input_ch=input_ch,
+        output_ch=output_ch,
+        skips=skips,
+        input_ch_views=input_ch_views,
+        use_viewdirs=args.use_viewdirs,
+    ).to(device)
+    grad_vars = list(model.parameters())
+
+    model_fine = None
+    if args.N_importance > 0:
+        model_fine = NeRF(
+            D=args.netdepth_fine,
+            W=args.netwidth_fine,
+            input_ch=input_ch,
+            output_ch=output_ch,
+            skips=skips,
+            input_ch_views=input_ch_views,
+            use_viewdirs=args.use_viewdirs,
+        ).to(device)
+        grad_vars += list(model_fine.parameters())
+
+    network_query_fn = lambda inputs, viewdirs, network_fn: run_network(
+        inputs,
+        viewdirs,
+        network_fn,
+        embed_fn=embed_fn,
+        embeddirs_fn=embeddirs_fn,
+        netchunk=args.netchunk,
+    )
+
+    # Create optimizer
+    optimizer = torch.optim.Adam(params=grad_vars, lr=args.lrate, betas=(0.9, 0.999))
+
+    start = 0
+    basedir = args.basedir
+    expname = args.expname
+
+    ##########################
+
+    # Load checkpoints
+    if args.ft_path is not None and args.ft_path != "None":
+        ckpts = [args.ft_path]
+    else:
+        ckpts = [
+            os.path.join(basedir, expname, f)
+            for f in sorted(os.listdir(os.path.join(basedir, expname)))
+            if "tar" in f
+        ]
+
+    print("Found ckpts", ckpts)
+    if len(ckpts) > 0 and not args.no_reload:
+        ckpt_path = ckpts[-1]
+        print("Reloading from", ckpt_path)
+        ckpt = torch.load(ckpt_path)
+
+        start = ckpt["global_step"]
+        optimizer.load_state_dict(ckpt["optimizer_state_dict"])
+
+        # Load model
+        model.load_state_dict(ckpt["network_fn_state_dict"])
+        if model_fine is not None:
+            model_fine.load_state_dict(ckpt["network_fine_state_dict"])
+
+    ##########################
+
+    render_kwargs_train = {
+        "network_query_fn": network_query_fn,
+        "perturb": args.perturb,
+        "N_importance": args.N_importance,
+        "network_fine": model_fine,
+        "N_samples": args.N_samples,
+        "network_fn": model,
+        "use_viewdirs": args.use_viewdirs,
+        "white_bkgd": args.white_bkgd,
+        "raw_noise_std": args.raw_noise_std,
+    }
+
+    # NDC only good for LLFF-style forward facing data
+    if args.dataset_type != "llff" or args.no_ndc:
+        print("Not ndc!")
+        render_kwargs_train["ndc"] = False
+        render_kwargs_train["lindisp"] = args.lindisp
+
+    render_kwargs_test = {k: render_kwargs_train[k] for k in render_kwargs_train}
+    render_kwargs_test["perturb"] = False
+    render_kwargs_test["raw_noise_std"] = 0.0
+
+    return render_kwargs_train, render_kwargs_test, start, grad_vars, optimizer
